requirements.txt
Flask
Werkzeug
mysql-connector-python
PyJWT
bcrypt
Flask-Mail
Flask-Caching
Pillow
python-dotenv
email-validator
gunicorn
redis
celery
requests
Jinja2
MarkupSafe
itsdangerous
click
blinker
importlib-metadata
zipp